# expression: implement vectorized evaluation for 'builtinHourSig'