@rogaps 2019-10-15
I am sorry. This implementation doesn't seem right. The test fails if negative durations are included in the test. I will create another PR to fix this.